add pagination in reviews

+++ mandiri-ios/Data/Responses/ReviewResponse.swift
@@ -41,7 +41,7 @@ struct ResultReview: Codable {
 struct AuthorDetails: Codable {
     let name, username: String?
     let avatarPath: String?
-    let rating: Int?
+    let rating: Double?
     
     enum CodingKeys: String, CodingKey {
         case name, username
